requirements.txt
langchain-openai
python-dotenv
pillow
